- Updated split utxo logic
