+++ lib/Wallet copy.js
@@ -301,6 +301,7 @@ class Wallet {
                 };
             }
             var balanceOfValidUTXOS = sumUTXOs(validUTXOs);
+            ///////
             // get all  of the UTXOSs above the threshold  of the wallet
             var UTXOSofWalletAboveThreshold = utxos.filter((utxo) => {
                 const value = new bignumber_js_1.BigNumber(utxo.value);
@@ -370,7 +371,9 @@ class Wallet {
             function filterUtxos(utxos, utxoThreshold) {
                 return utxos.filter((utxo) => {
                     const value = new bignumber_js_1.BigNumber(utxo.value);
-                    if (value.gt(new bignumber_js_1.BigNumber(25).times(1e6)) &&
+                    if (
+                    //  if (value.gt(new BigNumber(4).times(1e6))) {
+                    value.gt(new bignumber_js_1.BigNumber(25).times(1e6)) &&
                         value.lt(new bignumber_js_1.BigNumber(utxoThreshold).times(1e8))) {
                         return true;
                     }
@@ -398,152 +401,12 @@ class Wallet {
             }
         });
     }
-    splitUTXOS(utxos, keyPair, feeRate, utxoMinValue, UTXO_MIN_VALUE //UTXO_MIN_VALUE
-    ) {
-        return __awaiter(this, void 0, void 0, function* () {
-            // Filter the utxos so that they don't have value less than 0.04 HYDRA or 100
-            var validUTXOs = filterUtxos(utxos, UTXO_MIN_VALUE);
-            if (validUTXOs.length == 0) {
-                return {
-                    hex: "",
-                    error: "No UTXOs to optimize.",
-                };
-            }
-            // Sort them by size so it optimize the small ones first
-            validUTXOs.sort((lhs, rhs) => {
-                const lhsValue = new bignumber_js_1.BigNumber(lhs.value);
-                const rhsValue = new bignumber_js_1.BigNumber(rhs.value);
-                return lhsValue.minus(rhsValue).toNumber();
-            });
-            // Calculate the total balance
-            var balance = sumUTXOs(validUTXOs);
-            if (balance.lte(new bignumber_js_1.BigNumber(UTXO_MIN_VALUE).times(1e8))) {
-                return {
-                    hex: "",
-                    error: "Not enough balance for minimum UTXO.",
-                };
-            }
-            var fee;
-            var outputs;
-            // Calculate the total outputs with value of UTXO_MIN_VALUE
-            var totalOutputs = balance.dividedToIntegerBy(new bignumber_js_1.BigNumber(UTXO_MIN_VALUE).times(1e8));
-            if (totalOutputs.toNumber() > 100) {
-                // If more than 100 => therefore the balance is over 10000 => 100 outputs with 100 and 1 with the change
-                outputs = 101;
-                validUTXOs = selectUTXOs(validUTXOs, feeRate, new bignumber_js_1.BigNumber(UTXO_MIN_VALUE).times(1e8).times(100).toNumber(), outputs);
-                console.log(validUTXOs.length);
-                balance = sumUTXOs(validUTXOs);
-            }
-            else {
-                outputs = totalOutputs.toNumber();
-            }
-            var tx = new bitcoinjs_lib_1.TransactionBuilder(keyPair.network);
-            // Get the walet address
-            var from = this.address;
-            //Calculate the value for the transaction
-            var value = new bignumber_js_1.BigNumber(UTXO_MIN_VALUE).times(1e8).times(outputs);
-            // Calculate fee with the current inputs and outputs
-            fee = calculateFee(validUTXOs, outputs, feeRate, keyPair);
-            console.log(validUTXOs.length);
-            console.log(fee.toNumber());
-            if (fee.gt(balance)) {
-                return {
-                    hex: "",
-                    error: "Not enough balance to pay fee.",
-                };
-            }
-            //Remove one output if not enough balance for value + fee
-            if (balance.minus(fee).lt(value)) {
-                outputs -= 1;
-                value = value.minus(new bignumber_js_1.BigNumber(UTXO_MIN_VALUE).times(1e8));
-            }
-            // If after the substitution the value is less than the minimum utxo value no point for transaction
-            if (value.lt(new bignumber_js_1.BigNumber(UTXO_MIN_VALUE).times(1e8))) {
-                return {
-                    hex: "",
-                    error: "No enough balance for minimum UTXO.",
-                };
-            }
-            // Add the inputs
-            for (var i = 0; i < validUTXOs.length; i++) {
-                tx.addInput(validUTXOs[i].hash, validUTXOs[i].outputIndex);
-            }
-            // Add the outputs
-            tx.addOutput(from, balance
-                .minus(fee)
-                .minus(new bignumber_js_1.BigNumber(outputs - 1).times(UTXO_MIN_VALUE).times(1e8))
-                .toNumber());
-            for (var i = 0; i < outputs - 1; i++) {
-                tx.addOutput(from, new bignumber_js_1.BigNumber(UTXO_MIN_VALUE).times(1e8).toNumber());
-            }
-            // Sign the inputs
-            for (var i = 0; i < validUTXOs.length; i++) {
-                tx.sign(i, keyPair);
-            }
-            return {
-                hex: tx.build().toHex(),
-                error: "",
-            };
-            //////////////////SUPPORT FUNCTIONS/////////////////////
-            function filterUtxos(utxos, UTXO_MIN_VALUE) {
-                return utxos.filter((utxo) => {
-                    const value = new bignumber_js_1.BigNumber(utxo.value);
-                    if (value.gt(new bignumber_js_1.BigNumber(4).times(1e6))) {
-                        if (!value.eq(new bignumber_js_1.BigNumber(UTXO_MIN_VALUE).times(1e8))) {
-                            return true;
-                        }
-                    }
-                    return false;
-                });
-            }
-            //////
-            function sumUTXOs(utxos) {
-                let sum = new bignumber_js_1.BigNumber(0);
-                for (let utxo of utxos) {
-                    sum = sum.plus(utxo.value);
-                }
-                return sum;
-            }
-            /////////
-            function selectUTXOs(utxos, value, bytefee, outputs = 101) {
-                var totalValue = new bignumber_js_1.BigNumber(0);
-                var selected = [];
-                for (let utxo of utxos) {
-                    totalValue = totalValue.plus(utxo.value);
-                    selected.push(utxo);
-                    var txsize = selected.length * 102 + outputs * 31 + 10;
-                    if (totalValue.gt(txsize * bytefee + value)) {
-                        break;
-                    }
-                }
-                return selected;
-            }
-            ////////
-            function calculateFee(inputs, outputs, feeRate, keyPair) {
-                var tx = new bitcoinjs_lib_1.TransactionBuilder(keyPair.getNetwork());
-                for (var i = 0; i < inputs.length; i++) {
-                    ///adds the inputs to the tx
-                    tx.addInput(inputs[i].hash, inputs[i].outputIndex);
-                }
-                for (var i = 0; i <= outputs; i++) {
-                    tx.addOutput(from, new bignumber_js_1.BigNumber(UTXO_MIN_VALUE).times(1e8).toNumber());
-                }
-                // Sign the inputs
-                for (var i = 0; i < inputs.length; i++) {
-                    tx.sign(i, keyPair);
-                }
-                return new bignumber_js_1.BigNumber(tx.build().toHex().length).times(feeRate);
-            }
-            //////////////////SUPPORT FUNCTIONS/////////////////////
-        });
-    }
-    optimizeWalletUTXOS(utxoMinValue, utxoThreshold //UTXO_MIN_VALUE
-    ) {
+    optimizeWalletUTXOS(utxoMinValue, utxoThreshold) {
         return __awaiter(this, void 0, void 0, function* () {
             const utxos = yield this.getBitcoinjsUTXOs();
             const infoRes = yield this.insight.getBlockchainInfo();
             const feeRate = Math.ceil(infoRes.feeRate * 1e5);
-            let txResponse = yield this.splitUTXOS(utxos, this.keyPair, feeRate, utxoMinValue || 100, utxoThreshold || 150);
+            let txResponse = yield this.optimizeUTXOS(utxos, this.keyPair, feeRate, utxoMinValue || 100, utxoThreshold || 150);
             return txResponse.hex !== ""
                 ? yield this.sendRawTx(txResponse.hex)
                 : txResponse.error;
@@ -554,4 +417,4 @@ exports.Wallet = Wallet;
 // TODO
 // hrc20 lookup
 // estimateCall
-//# sourceMappingURL=Wallet.js.map
+//# sourceMappingURL=Wallet copy.js.map
