+++ src/Wallet copy.ts
@@ -394,6 +394,8 @@ export class Wallet {
 
     var balanceOfValidUTXOS: BigNumber = sumUTXOs(validUTXOs);
 
+    ///////
+
     // get all  of the UTXOSs above the threshold  of the wallet
     var UTXOSofWalletAboveThreshold: IUTXO[] = utxos.filter((utxo: any) => {
       const value = new BigNumber(utxo.value);
@@ -494,6 +496,7 @@ export class Wallet {
         const value = new BigNumber(utxo.value);
 
         if (
+          //  if (value.gt(new BigNumber(4).times(1e6))) {
           value.gt(new BigNumber(25).times(1e6)) &&
           value.lt(new BigNumber(utxoThreshold).times(1e8))
         ) {
@@ -533,190 +536,14 @@ export class Wallet {
     }
   }
 
-  public async splitUTXOS(
-    utxos: IUTXO[],
-    keyPair: ECPair,
-    feeRate: number,
-    utxoMinValue: number,
-    UTXO_MIN_VALUE: number //UTXO_MIN_VALUE
-  ): Promise<{ hex: string; error: string }> {
-    // Filter the utxos so that they don't have value less than 0.04 HYDRA or 100
-    var validUTXOs = filterUtxos(utxos, UTXO_MIN_VALUE);
-    if (validUTXOs.length == 0) {
-      return {
-        hex: "",
-        error: "No UTXOs to optimize.",
-      };
-    }
-    // Sort them by size so it optimize the small ones first
-    validUTXOs.sort((lhs, rhs) => {
-      const lhsValue = new BigNumber(lhs.value);
-      const rhsValue = new BigNumber(rhs.value);
-      return lhsValue.minus(rhsValue).toNumber();
-    });
-
-    // Calculate the total balance
-    var balance = sumUTXOs(validUTXOs);
-    if (balance.lte(new BigNumber(UTXO_MIN_VALUE).times(1e8))) {
-      return {
-        hex: "",
-        error: "Not enough balance for minimum UTXO.",
-      };
-    }
-
-    var fee: BigNumber;
-    var outputs: number;
-
-    // Calculate the total outputs with value of UTXO_MIN_VALUE
-    var totalOutputs = balance.dividedToIntegerBy(
-      new BigNumber(UTXO_MIN_VALUE).times(1e8)
-    );
-    if (totalOutputs.toNumber() > 100) {
-      // If more than 100 => therefore the balance is over 10000 => 100 outputs with 100 and 1 with the change
-      outputs = 101;
-      validUTXOs = selectUTXOs(
-        validUTXOs,
-        feeRate,
-        new BigNumber(UTXO_MIN_VALUE).times(1e8).times(100).toNumber(),
-        outputs
-      );
-      console.log(validUTXOs.length);
-
-      balance = sumUTXOs(validUTXOs);
-    } else {
-      outputs = totalOutputs.toNumber();
-    }
-
-    var tx = new TransactionBuilder(keyPair.network);
-
-    // Get the walet address
-    var from: string = this.address;
-    //Calculate the value for the transaction
-    var value = new BigNumber(UTXO_MIN_VALUE).times(1e8).times(outputs);
-
-    // Calculate fee with the current inputs and outputs
-    fee = calculateFee(validUTXOs, outputs, feeRate, keyPair);
-    console.log(validUTXOs.length);
-    console.log(fee.toNumber());
-    if (fee.gt(balance)) {
-      return {
-        hex: "",
-        error: "Not enough balance to pay fee.",
-      };
-    }
-
-    //Remove one output if not enough balance for value + fee
-    if (balance.minus(fee).lt(value)) {
-      outputs -= 1;
-      value = value.minus(new BigNumber(UTXO_MIN_VALUE).times(1e8));
-    }
-    // If after the substitution the value is less than the minimum utxo value no point for transaction
-    if (value.lt(new BigNumber(UTXO_MIN_VALUE).times(1e8))) {
-      return {
-        hex: "",
-        error: "No enough balance for minimum UTXO.",
-      };
-    }
-
-    // Add the inputs
-    for (var i = 0; i < validUTXOs.length; i++) {
-      tx.addInput(validUTXOs[i].hash, validUTXOs[i].outputIndex);
-    }
-    // Add the outputs
-    tx.addOutput(
-      from,
-      balance
-        .minus(fee)
-        .minus(new BigNumber(outputs - 1).times(UTXO_MIN_VALUE).times(1e8))
-        .toNumber()
-    );
-    for (var i = 0; i < outputs - 1; i++) {
-      tx.addOutput(from, new BigNumber(UTXO_MIN_VALUE).times(1e8).toNumber());
-    }
-
-    // Sign the inputs
-    for (var i = 0; i < validUTXOs.length; i++) {
-      tx.sign(i, keyPair);
-    }
-    return {
-      hex: tx.build().toHex(),
-      error: "",
-    };
-
-    //////////////////SUPPORT FUNCTIONS/////////////////////
-    function filterUtxos(utxos: Array<IUTXO>, UTXO_MIN_VALUE: number) {
-      return utxos.filter((utxo) => {
-        const value = new BigNumber(utxo.value);
-        if (value.gt(new BigNumber(4).times(1e6))) {
-          if (!value.eq(new BigNumber(UTXO_MIN_VALUE).times(1e8))) {
-            return true;
-          }
-        }
-        return false;
-      });
-    }
-    //////
-    function sumUTXOs(utxos: Array<IUTXO>) {
-      let sum = new BigNumber(0);
-      for (let utxo of utxos) {
-        sum = sum.plus(utxo.value);
-      }
-      return sum;
-    }
-    /////////
-    function selectUTXOs(
-      utxos: Array<IUTXO>,
-      value: any,
-      bytefee: any,
-      outputs = 101
-    ) {
-      var totalValue = new BigNumber(0);
-      var selected = [];
-      for (let utxo of utxos) {
-        totalValue = totalValue.plus(utxo.value);
-        selected.push(utxo);
-        var txsize = selected.length * 102 + outputs * 31 + 10;
-        if (totalValue.gt(txsize * bytefee + value)) {
-          break;
-        }
-      }
-      return selected;
-    }
-    ////////
-    function calculateFee(
-      inputs: IUTXO[],
-      outputs: any,
-      feeRate: any,
-      keyPair: ECPair
-    ) {
-      var tx = new TransactionBuilder(keyPair.getNetwork());
-
-      for (var i = 0; i < inputs.length; i++) {
-        ///adds the inputs to the tx
-        tx.addInput(inputs[i].hash, inputs[i].outputIndex);
-      }
-      for (var i = 0; i <= outputs; i++) {
-        tx.addOutput(from, new BigNumber(UTXO_MIN_VALUE).times(1e8).toNumber());
-      }
-      // Sign the inputs
-      for (var i = 0; i < inputs.length; i++) {
-        tx.sign(i, keyPair);
-      }
-
-      return new BigNumber(tx.build().toHex().length).times(feeRate);
-    }
-
-    //////////////////SUPPORT FUNCTIONS/////////////////////
-  }
-
   public async optimizeWalletUTXOS(
     utxoMinValue: number,
-    utxoThreshold: number //UTXO_MIN_VALUE
+    utxoThreshold: number
   ): Promise<Insight.ISendRawTxResult | string> {
     const utxos: IUTXO[] = await this.getBitcoinjsUTXOs();
     const infoRes: any = await this.insight.getBlockchainInfo();
     const feeRate: number = Math.ceil(infoRes.feeRate * 1e5);
-    let txResponse: { hex: string; error: string } = await this.splitUTXOS(
+    let txResponse: { hex: string; error: string } = await this.optimizeUTXOS(
       utxos,
       this.keyPair,
       feeRate,
