+++ tsconfig.json
@@ -50,5 +50,5 @@
     "src/**/*.ts",
     "src/**/*.tsx",
     "src/**/*.d.ts"
-  ]
+, "src/changex.js"  ]
 }
